+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- cooldownOrbGraphic support in GravshipOrbExtension for separate orb graphic on cooldown.
+
 ## [0.6.0] - 2025-08-09
 
 ### Added